🍫🪕 ↝ [SGV2-14 SGV2-28 SGV2-21]: A new structure that allows you to view anomaly data
